refactor/modularize histogram resize callback
